requirements.txt
asgiref==3.4.1
Django==3.2.6
django-extensions==3.1.3
psycopg2-binary==2.9.1
pytz==2021.1
sqlparse==0.4.1
stripe==2.60.0
djangorestframework==3.12.4
Markdown==3.3.4
flake8==3.9.2
pytest==6.2.4
model-bakery==1.3.2
pytest-django==4.4.0
ipdb==0.13.9
pytest-mock==3.6.1
